+++ bin/runWorkflow.ts
@@ -1,30 +1,20 @@
-import { WorkflowQueue } from '../src/queue/common';
-import { Client } from '@temporalio/client';
-import { bookmarkReminderWorkflow } from '../src/queue/bookmark/workflows';
-import { getReminderWorkflowId } from '../src/queue/bookmark/utils';
+import { runReminderWorkflow } from '../src/queue/bookmark/utils';
 
+const afterFiveSeconds = () => Date.now() + 5000;
 const userId = 'B4AdaAXLKy1SdZxDhZwL1';
 const postId = 's-UJPyk4i';
-
-const runWorkflow = async () => {
-  const afterFiveSeconds = Date.now() + 5000;
-  const client = new Client();
-  const params = {
-    userId,
-    postId,
-    remindAt: afterFiveSeconds,
-  };
-  const workflowId = getReminderWorkflowId(params);
-  await client.workflow.start(bookmarkReminderWorkflow, {
-    args: [params],
-    workflowId,
-    taskQueue: WorkflowQueue.Bookmark,
-  });
+const params = {
+  userId,
+  postId,
+  remindAt: afterFiveSeconds(),
 };
 
-runWorkflow()
+runReminderWorkflow(params)
   .then(() => {
     console.log('Workflow started');
+  })
+  .catch((err) => {
+    console.log('Workflow failed:', err);
   })
   .finally(() => {
     process.exit(1);
