+++ src/queue/bookmark/utils.ts
@@ -1,15 +1,29 @@
 import {
   generateWorkflowId,
+  WorkflowQueue,
   WorkflowTopic,
   WorkflowTopicScope,
 } from '../common';
-import { BookmarkReminderParams } from './workflows';
+import { BookmarkReminderParams, bookmarkReminderWorkflow } from './workflows';
+import { getTemporalClient } from '../client';
 
 export const getReminderWorkflowId = ({
   userId,
   postId,
-}: Pick<BookmarkReminderParams, 'userId' | 'postId'>) =>
+  remindAt,
+}: BookmarkReminderParams) =>
   generateWorkflowId(WorkflowTopic.Bookmark, WorkflowTopicScope.Reminder, [
     userId,
     postId,
+    remindAt.toString(),
   ]);
+
+export const runReminderWorkflow = async (params: BookmarkReminderParams) => {
+  const workflowId = getReminderWorkflowId(params);
+  const client = await getTemporalClient();
+  client.workflow.start(bookmarkReminderWorkflow, {
+    args: [params],
+    workflowId,
+    taskQueue: WorkflowQueue.Bookmark,
+  });
+};
